Build wrapped_clang and wrapped_clang_pp in action
This compiles wrapped_clang and wrapped_clang_pp inside of an action.
https://github.com/bazelbuild/bazel/issues/8902 has a larger definition
of the issue.

This action needs the DEVELOPER_DIR as an input in order to select the
correct clang. This is passed in from apple_common.XcodeVersionConfig.

Note that this works for local execution and needs to address the TODO
for cc_wrapper.sh in tools/cpp/osx_cc_configure.bzl.